move axle unit helpers into separate file
