+++ frontend/src/features/permits/helpers/serialize/serializePermitVehicleConfiguration.ts
@@ -1,21 +1,7 @@
+import { mergeInteraxleSpacing } from "../../../../common/helpers/axleUnitHelper";
 import { convertToNumberIfValid } from "../../../../common/helpers/numeric/convertToNumberIfValid";
-import { AxleUnit } from "../../../../common/types/AxleUnit";
 import { Nullable, RequiredOrNull } from "../../../../common/types/common";
 import { PermitVehicleConfiguration } from "../../types/PermitVehicleConfiguration";
-
-/** Removes individual interaxle spacing rows and merges their value into the next axleConfiguration object */
-const mergeInteraxleSpacingRows = (
-  axleConfiguration: AxleUnit[],
-  startIndex: number,
-) => {
-  const merged = [...axleConfiguration];
-  for (let i = startIndex; i < merged.length - 1; i++) {
-    merged[i + 1].interaxleSpacing = merged[i].interaxleSpacing;
-    merged.splice(i, 1);
-  }
-
-  return merged;
-};
 
 /**
  * Serialize permit vehicle configuration values as request payload.
@@ -31,7 +17,7 @@ export const serializePermitVehicleConfiguration = (
           ? vehicleConfiguration.trailers.map((trailer) => ({
               ...trailer,
               axleConfiguration: trailer.axleConfiguration
-                ? mergeInteraxleSpacingRows([...trailer.axleConfiguration], 0)
+                ? mergeInteraxleSpacing([...trailer.axleConfiguration], 0)
                 : null,
             }))
           : null,
@@ -58,7 +44,7 @@ export const serializePermitVehicleConfiguration = (
         loadedGVW: convertToNumberIfValid(vehicleConfiguration.loadedGVW, null),
         netWeight: convertToNumberIfValid(vehicleConfiguration.netWeight, null),
         axleConfiguration: vehicleConfiguration.axleConfiguration
-          ? mergeInteraxleSpacingRows(
+          ? mergeInteraxleSpacing(
               [...vehicleConfiguration.axleConfiguration],
               1,
             )
